rename package to make sure the samples are using public interfaces
