+++ src/main/resources/samples.conf
@@ -32,7 +32,7 @@ dataObjects {
   custom-my-df {
     type = CustomDfDataObject
     creator {
-      class-name = io.smartdatalake.samples.CreateMyDf
+      class-name = com.test.CreateMyDf
     }
   }
 
@@ -116,7 +116,7 @@ dataObjects {
   custom-rating-df {
     type = CustomDfDataObject
     creator {
-      class-name = io.smartdatalake.samples.CreateRatingDf
+      class-name = com.test.CreateRatingDf
     }
   }
 
@@ -206,7 +206,7 @@ actions {
     type = CopyAction
     inputId = ab-csv-org
     outputId = ab-reduced-hsqldb
-    transformer.class-name = io.smartdatalake.samples.ReduceNycCSVTransformer
+    transformer.class-name = com.test.ReduceNycCSVTransformer
     metadata {
       feed = ab-jdbc
     }
@@ -241,7 +241,7 @@ actions {
     type = CustomSparkAction
     inputIds = [custom-rating-csv1, custom-rating-csv2]
     outputIds = [custom-rating-csv-agg]
-    transformer.class-name = io.smartdatalake.samples.RatingTransformer
+    transformer.class-name = com.test.RatingTransformer
     metadata {
       feed = custom-rating-csv
     }
